提交代码

+++ internal/model/tag.go
@@ -8,7 +8,7 @@ import (
 type Tag struct {
 	*Model
 	Name		string 	`json:"name"`
-	State		uint32	`json:"state"`
+	State		uint8	`json:"state"`
 }
 
 type TagSwagger struct {
